added logic for amount/percentage changes for sections
What's new:
clicking on a row selects it and gives user option to update it (if a section has not been chosen it defaults to an existing section- not fully implemented as it always assumes a "savings" section to exist)

radio buttons indicate whether input should be treated as a percentage or an amount

the appropriate section property is set accordingly.

The selected table row is then updated with the new data

What's Changed:

there is now a function to handle populating a row with section data (populateRow in dynamicTable.kt)
this is now used both for generating the TextViews that hold the data in each 'table cell' when the table is first generated and updating the data in the existing table cells when needed.

the round function in Budget.kt is now outside the class entirely. Not the best solution lol but I needed to get to it in dynamicTable.kt and I was lazy.